requirements.txt
streamlit
pandas
plotly
supabase
python-dateutil